Fixed float vs int problem

+++ direct/src/directtools/DirectGrid.py
@@ -75,7 +75,7 @@ class DirectGrid(NodePath,PandaObject):
         self.centerLines.reset()
 
         # Now redraw lines
-        numLines = math.ceil(self.gridSize/self.gridSpacing)
+        numLines = int(math.ceil(self.gridSize/self.gridSpacing))
         scaledSize = numLines * self.gridSpacing
  
         center = self.centerLines
